+++ R/simulate_infections.R
@@ -164,7 +164,7 @@ simulate_infections <- function(R,
         "!" = "Cannot simulate from estimated nonparametric
         delays.",
         "i" = "Use a fixed {.fn NonParametric} distribution
-        instead of {.fn EstimatedNonParametric}."
+        instead of {.fn Dirichlet}."
       )
     )
   }